ipc-rag-predictor/requirements.txt



### 2. requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9

# PDF Processing
pymupdf==1.24.10
pdfplumber==0.11.4

# Embeddings & Vector Store
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy==1.26.4

# MLflow
mlflow==2.16.2

# Utils
python-dotenv==1.0.1
structlog==24.4.0
rich==13.8.1

# CORS & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# Async support
anyio==4.4.0